eal: fix 32-bit structure marker
Use the 32bit typedef for structure 4B alignment marking.

Fixes: 2b393160a480 ("eal: introduce structure markers")
Cc: stable@dpdk.org

Acked-by: Xiaolong Ye <xiaolong.ye@intel.com>
Acked-by: Jerin Jacob <jerinj@marvell.com>

+++ lib/librte_eal/common/include/rte_common.h
@@ -344,7 +344,7 @@ __extension__ typedef uint8_t  RTE_MARKER8[0];
 /** Marker for 2B alignment in a structure. */
 __extension__ typedef uint16_t RTE_MARKER16[0];
 /** Marker for 4B alignment in a structure. */
-__extension__ typedef uint16_t RTE_MARKER32[0];
+__extension__ typedef uint32_t RTE_MARKER32[0];
 /** Marker for 8B alignment in a structure. */
 __extension__ typedef uint64_t RTE_MARKER64[0];
